Removed the deprecated sunpy.sun.sun module and the deprecated sunpy.coordinates.ephemeris functions
