+++ docs/code_ref/sun.rst
@@ -5,8 +5,6 @@ SunPy sun
 
 The sun submodule contains constants, parameters and models of the Sun.
 
-.. automodapi:: sunpy.sun.sun
-
 .. automodapi:: sunpy.sun.constants
 
 .. automodapi:: sunpy.sun.models
