+++ docs/dev_guide/sunpy_stability.yaml
@@ -30,7 +30,7 @@ roi:
   comments: There is very little functionality.
 sun:
   status: dev
-  comments: Sub-module `constants` is stable.  Sub-module `models` will likely have major changes.  Sub-module `sun` is deprecated in 1.0 and will be removed in 2.0.
+  comments: Sub-module `constants` is stable.  Sub-module `models` will likely have major changes.
 time:
   status: stable
   comments: Any future changes should largely maintain backwards compatibility.
