+++ sunpy/instr/goes.py
@@ -61,7 +61,7 @@ import astropy.units as u
 from astropy.time import TimeDelta
 
 from sunpy import timeseries
-from sunpy.coordinates import get_sunearth_distance, sun
+from sunpy.coordinates import sun
 from sunpy.data import manager
 from sunpy.sun import constants
 from sunpy.time import parse_time
